implement max-runtime and make it really work

+++ libswfdec-gtk/swfdec_gtk_player.c
@@ -174,7 +174,7 @@ swfdec_gtk_player_new (SwfdecAsDebugger *debugger)
 
   swfdec_init ();
   player = g_object_new (SWFDEC_TYPE_GTK_PLAYER, "debugger", debugger, 
-      "max-runtime", 10 * 1000, NULL);
+      "max-runtime", (gulong) 10 * 1000, NULL);
 
   return player;
 }
